Added support to check minimum required version to show edit file attachments
